Data Parsing crash Issue Fix

+++ app/src/main/java/com/afss/impresario/Services/DataService.java
@@ -13,9 +13,9 @@ public class DataService {
         {
             if (txnTypeList.get(i).contains("exp"))
             {
-                balance-=Integer.parseInt(txnAmountList.get(i));
+                balance-=Double.parseDouble(txnAmountList.get(i));
             }else {
-                balance+=Integer.parseInt(txnAmountList.get(i));
+                balance+=Double.parseDouble(txnAmountList.get(i));
             }
 
         }
